backend/requirements.txt
anthropic
fastapi
uvicorn
python-multipart
pypdf
sentence-transformers
faiss-cpu
httpx
streamlit
python-dotenv
